Fix BCELoss + AMP: compute loss outside autocast
BCELoss is explicitly blocked by PyTorch autocast because
sigmoid outputs in float16 can cause log(0) precision issues.
Fix: run the forward pass in autocast (float16 for speed) but
compute the loss in float32 by casting model output with
out.float() before passing to criterion.

+++ engine_branch1.py
@@ -34,11 +34,11 @@ def train_one_epoch(train_loader,
         images = images.to(device, non_blocking=True).float()
         targets = targets.to(device, non_blocking=True).float()
 
-        # AMP: autocast runs forward pass and loss in float16 where safe,
-        # keeping numerically sensitive ops (e.g. BCELoss) in float32
+        # AMP: forward pass in float16 for speed, loss in float32 for safety
+        # (BCELoss is not autocast-safe due to sigmoid+log precision issues)
         with autocast(enabled=use_amp):
             out = model(images)
-            loss = criterion(out, targets)
+        loss = criterion(out.float(), targets)
 
         if use_amp:
             scaler.scale(loss).backward()
